requirements.txt
aiogram==3.15.0
httpx==0.28.1
python-dotenv==1.1.0
